caught runtime error for event boolean exp, updated visitor classes to inherit from ErrorVisitor as setup for the semantic visitor; NEED to fix function parse error!
